refactor(billing): break db/schema <-> schemas type cycle
Pre-existing circular dep: db/schema.ts types its columns from schema
types in schemas/index.ts, while schemas/index.ts imported
BillingDocumentSelect (only for the row->response mapper). Move
toBillingDocumentResponse into schemas/toResponse.ts so schemas/index.ts
no longer imports db/schema, leaving a clean one-way edge
(db/schema -> schemas/index). depcruise no-circular: 0. 88 tests pass.
